+++ src/app/store/header/menu/store-header-menu.component.ts
@@ -5,7 +5,7 @@ import { Observable, of } from 'rxjs';
 import { pluck, switchMap } from 'rxjs/operators';
 import { AppService } from 'src/app/app.service';
 import { ConfirmationDialogComponent, ConfirmationDialogData } from 'src/app/shared/components/confirmation-dialogconfirmation-dialog.component';
-import { EditProfileFormDialogComponent } from 'src/app/shared/edit-profile-form-dialog/edit-profile-form-dialog.component';
+import { EditProfileFormDialogComponent } from 'src/app/shared/components/edit-profile-form-dialog/edit-profile-form-dialog.component';
 import { LOGOUT_MESSAGE } from 'src/text/messages';
 
 @Component({
